created BlockedSiteCard component
