+++ client/routes.js
@@ -1,47 +1,47 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import { withRouter, Route, Switch, Redirect } from 'react-router-dom'
-import { Login, Signup } from './components/AuthForm'
-import SandboxLogin from './components/AuthForm/SandboxLogin'
-import CreateSession from './components/Timer/CreateSession'
-import Timer from './components/Timer'
-import Dashboard from './components/Dashboard'
-import { loadSessions } from './store/sessions'
-import { updateBlackList } from './store/blackList'
-import { loadBlocks } from './store/blocks'
-import { loadSites } from './store/sites'
-import { getSites } from './store/blockSites'
-import BlockError from './components/BlockError'
-import BlockSites from './components/BlockSites'
-import Friends from './components/Friends/Friends'
-import About from './components/About'
-import Home from './components/Home'
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import { withRouter, Route, Switch, Redirect } from "react-router-dom";
+import { Login, Signup } from "./components/AuthForm";
+import SandboxLogin from "./components/AuthForm/SandboxLogin";
+import CreateSession from "./components/Timer/CreateSession";
+import Timer from "./components/Timer";
+import Dashboard from "./components/Dashboard";
+import { loadSessions } from "./store/sessions";
+import { updateBlackList } from "./store/blackList";
+import { loadBlocks } from "./store/blocks";
+import { loadSites } from "./store/sites";
+import { getSites } from "./store/blockSites";
+import BlockError from "./components/BlockError";
+import BlockSites from "./components/BlockList";
+import Friends from "./components/Friends/Friends";
+import About from "./components/About";
+import Home from "./components/Home";
 
 class Routes extends Component {
   constructor(props) {
-    super(props)
+    super(props);
   }
   async componentDidMount() {
-    await this.props.loadInitialData()
+    await this.props.loadInitialData();
   }
   async componentDidUpdate(prevProps) {
     if (this.props.auth.id && this.props.auth.id !== prevProps.auth.id) {
-      await this.props.getSites(this.props.auth.id)
+      await this.props.getSites(this.props.auth.id);
     }
-    if (navigator.userAgent.indexOf('Chrome') !== -1) {
-      chrome?.runtime?.sendMessage('nmfhcdkehekkflbopjlnnihncpnlejho', {
-        message: 'set-blocked-sites',
+    if (navigator.userAgent.indexOf("Chrome") !== -1) {
+      chrome?.runtime?.sendMessage("nmfhcdkehekkflbopjlnnihncpnlejho", {
+        message: "set-blocked-sites",
 
         blockedSites: this.props.blockedSites.filter((each) => {
-          return each.blacklist.blockingEnabled === true
+          return each.blacklist.blockingEnabled === true;
         }),
         currUser: this.props.auth.id,
-      })
+      });
     }
   }
 
   render() {
-    const { isLoggedIn, auth, blackList, updateB, timer } = this.props
+    const { isLoggedIn, auth, blackList, updateB, timer } = this.props;
 
     return (
       <div>
@@ -79,7 +79,7 @@ class Routes extends Component {
           </Switch>
         )}
       </div>
-    )
+    );
   }
 }
 
@@ -90,25 +90,25 @@ const mapState = (state) => {
     auth: state.auth,
     blackList: state.blackList,
     blockedSites: state.blockedSites,
-  }
-}
+  };
+};
 
 const mapDispatch = (dispatch) => {
   return {
     loadInitialData() {
-      dispatch(loadSessions())
-      dispatch(loadSites())
-      dispatch(loadBlocks())
+      dispatch(loadSessions());
+      dispatch(loadSites());
+      dispatch(loadBlocks());
     },
 
     updateB: (blackListId, blackListInfo) => {
-      return dispatch(updateBlackList(blackListId, blackListInfo))
+      return dispatch(updateBlackList(blackListId, blackListInfo));
     },
 
     getSites: (userId) => {
-      return dispatch(getSites(userId))
+      return dispatch(getSites(userId));
     },
-  }
-}
+  };
+};
 
-export default withRouter(connect(mapState, mapDispatch)(Routes))
+export default withRouter(connect(mapState, mapDispatch)(Routes));
